+++ controllers/categoryController.js
@@ -5,6 +5,7 @@ const Category = require("../models/category");
 const Item = require("../models/item");
 const brand = require("../models/brand");
 const category = require("../models/category");
+const password_to_match = require("../config/adminPassword");
 
 exports.category_list = asyncHandler(async (req, res, next) => {
   const all_categories = await Category.find({}).sort({ name: 1 }).exec();
@@ -46,11 +47,13 @@ exports.category_create_post = [
     .trim()
     .isLength({ min: 10 })
     .escape(),
-
+  body("adminpassword", "Admin password must not be empty")
+    .trim()
+    .isLength({ min: 1 })
+    .escape(),
   // Process request after validation and sanitization.
   asyncHandler(async (req, res, next) => {
-    // Extract the validation errors from a request.
-    const errors = validationResult(req);
+    const entered_password = req.body.adminpassword;
 
     //Create a new Category object with the sanitized name and description
     const category = new Category({
@@ -58,27 +61,38 @@ exports.category_create_post = [
       description: req.body.description,
     });
 
-    if (!errors.isEmpty()) {
-      // There are errors. Render the form again with sanitized values/error messages.
+    if (entered_password === password_to_match) {
+      // Extract the validation errors from a request.
+      const errors = validationResult(req);
+
+      if (!errors.isEmpty()) {
+        // There are errors. Render the form again with sanitized values/error messages.
+        res.render("category_form", {
+          title: "Create Category",
+          category: category,
+          errors: errors.array(),
+        });
+        return;
+      } else {
+        // Data from form is valid.
+        // Check if Category with same name already exists.
+        const categoryExists = await Category.findOne({ name: req.body.name })
+          .collation({ locale: "en", strength: 2 })
+          .exec();
+        if (categoryExists) {
+          //Category exists. Redirect to its details page
+          res.redirect(categoryExists.url);
+        } else {
+          await category.save();
+          res.redirect(category.url);
+        }
+      }
+    } else {
       res.render("category_form", {
         title: "Create Category",
         category: category,
-        errors: errors.array(),
+        errors: [{ msg: "Password is incorrect" }],
       });
-      return;
-    } else {
-      // Data from form is valid.
-      // Check if Category with same name already exists.
-      const categoryExists = await Category.findOne({ name: req.body.name })
-        .collation({ locale: "en", strength: 2 })
-        .exec();
-      if (categoryExists) {
-        //Category exists. Redirect to its details page
-        res.redirect(categoryExists.url);
-      } else {
-        await category.save();
-        res.redirect(category.url);
-      }
     }
   }),
 ];
@@ -99,26 +113,45 @@ exports.category_delete_get = asyncHandler(async (req, res, next) => {
     items_in_category: items_in_category,
   });
 });
-exports.category_delete_post = asyncHandler(async (req, res, next) => {
-  const [category, items_in_category] = await Promise.all([
-    Category.findById(req.params.id).exec(),
-    Item.find({ category: req.params.id }, "name description")
-      .sort({ name: 1 })
-      .exec(),
-  ]);
-  if (items_in_category.length > 0) {
-    //Category has items. Render in the same way as for GET
-    res.render("category_delete", {
-      title: "Delete Category",
-      category: category,
-      items_in_category: items_in_category,
-    });
-    return;
-  } else {
-    await Category.findByIdAndDelete(req.params.id);
-    res.redirect("/catalog/categories");
-  }
-});
+exports.category_delete_post = [
+  body("adminpassword", "Admin password must not be empty.")
+    .trim()
+    .isLength({ min: 1 })
+    .escape(),
+
+  asyncHandler(async (req, res, next) => {
+    const entered_password = req.body.adminpassword;
+    const [category, items_in_category] = await Promise.all([
+      Category.findById(req.params.id).exec(),
+      Item.find({ category: req.params.id }, "name description")
+        .sort({ name: 1 })
+        .exec(),
+    ]);
+    if (entered_password === password_to_match) {
+      if (items_in_category.length > 0) {
+        //Category has items. Render in the same way as for GET
+        res.render("category_delete", {
+          title: "Delete Category",
+          category: category,
+          items_in_category: items_in_category,
+        });
+        return;
+      } else {
+        await Category.findByIdAndDelete(req.params.id);
+        res.redirect("/catalog/categories");
+      }
+    } else {
+      const errors = ["Password is incorrect"];
+      res.render("category_delete", {
+        title: "Delete Category",
+        category: category,
+        items_in_category: items_in_category,
+        errors: errors,
+      });
+    }
+  }),
+];
+
 exports.category_update_get = asyncHandler(async (req, res, next) => {
   const category = await Category.findById(req.params.id);
   if (category === null) {
@@ -143,34 +176,49 @@ exports.category_update_post = [
     .trim()
     .isLength({ min: 10 })
     .escape(),
+  body("adminpassword", "Admin password must not be empty")
+    .trim()
+    .isLength({ min: 1 })
+    .escape(),
 
   // Process request after validation and sanitization.
   asyncHandler(async (req, res, next) => {
-    // Extract the validation errors from a request.
-    const errors = validationResult(req);
-
+    const entered_password = req.body.adminpassword;
     const category = new Category({
       name: req.body.name,
       description: req.body.description,
       _id: req.params
         .id /*// This is required, or a new ID will be assigned! */,
     });
-    if (!errors.isEmpty()) {
-      // There are errors. Render the form again with sanitized values/error messages.
+
+    if (entered_password === password_to_match) {
+      // Extract the validation errors from a request.
+      const errors = validationResult(req);
+
+      if (!errors.isEmpty()) {
+        // There are errors. Render the form again with sanitized values/error messages.
+        res.render("category_form", {
+          title: "Update Category",
+          category: category,
+          errors: errors.array(),
+        });
+        return;
+      } else {
+        // Data from form is valid.
+        const updatedCategory = await Category.findByIdAndUpdate(
+          req.params.id,
+          category,
+          {}
+        );
+        res.redirect(updatedCategory.url);
+      }
+    } else {
+      //password incorrect
       res.render("category_form", {
         title: "Update Category",
         category: category,
-        errors: errors.array(),
+        errors: [{ msg: "Password is incorrect" }],
       });
-      return;
-    } else {
-      // Data from form is valid.
-      const updatedCategory = await Category.findByIdAndUpdate(
-        req.params.id,
-        category,
-        {}
-      );
-      res.redirect(updatedCategory.url);
     }
   }),
 ];
